pyproject.toml
[project]
name = "fairrcheck"
version = "0.1.0"
description = "FAIRR compliance checker for HPC / computational research projects."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "reportlab>=4.0",
]

[project.scripts]
fairrcheck = "fairrcheck.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
